PIG-3117: A debug mode in which pig does not delete temporary files (ihadanny via cheolsoo)
git-svn-id: https://svn.apache.org/repos/asf/pig/trunk@1533693 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/MapReduceLauncher.java
@@ -415,9 +415,11 @@ public class MapReduceLauncher extends Launcher{
             failed = true;
         }
 
-        // Clean up all the intermediate data
-        for (String path : intermediateVisitor.getIntermediate()) {
-            FileLocalizer.delete(path, pc);
+        if (!"false".equalsIgnoreCase(pc.getProperties().getProperty(PigConfiguration.PIG_DELETE_TEMP_FILE))) {
+            // Clean up all the intermediate data
+            for (String path : intermediateVisitor.getIntermediate()) {
+                FileLocalizer.delete(path, pc);
+            }
         }
 
         // Look to see if any jobs failed.  If so, we need to report that.
